pyproject.toml
[project]
name = "poseidon-sim"
version = "0.0.0"
description = "POSEIDON MVP - maritime test and evaluation environment for AUV and SSV performance under controllable ocean and navigation conditions."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "TBD" }
authors = [
    { name = "POSEIDON MVP contributors" },
]
dependencies = []

[project.optional-dependencies]
# Dev-only tools (test runner, future linters) kept separate from the
# runtime extras so edge / Profile B installs never pull them.
dev = [
    "pytest>=8.0",
]
# Offline MCAP analysis (Tier-2 evaluation + contract-check CLI).
# Kept in an optional extra so mission-essential core has no pip deps
# (AGENTS.md Section 3, Profile B).
eval = [
    "mcap>=1.2.1",
    "mcap-ros2-support>=0.5.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# The repo is colcon-first; Python modules live under `poseidon-sim/`
# which hatchling cannot package because of the dash. Until a proper
# src-layout package lands (see `SYSTEM_DESIGN.md` Section 18), emit
# an empty wheel so `uv sync --extra eval` can install optional deps
# (mcap, mcap-ros2-support) without failing on project-build.
bypass-selection = true

[tool.uv]
# uv-specific settings. Keep dependency-groups empty until the first dep
# is adopted per OPEN_SOURCE_STACK.md Section 2.

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
    "build",
    "dist",
    ".venv",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
# Make the (dash-named) poseidon-sim source tree importable as
# `evaluation.metrics...` without installing it as a wheel. Hatchling
# cannot package a directory with a dash in its name, and the repo
# layout is colcon-first, so path injection is the cleanest option.
pythonpath = ["poseidon-sim"]
